Removed device_manager because it is moved into device_utils and adaptet the code to use the correct device_manager
